requirements.txt
arxiv
pandas
python-dotenv
requests
beautifulsoup4